+++ src/guzzle/src/RingPHP/CoroutineHandler.php
@@ -147,11 +147,16 @@ class CoroutineHandler
         $statusCode = $client->statusCode;
         $errCode = $client->errCode;
 
-        if ($statusCode === -1) {
+        if ($statusCode === SWOOLE_HTTP_CLIENT_ESTATUS_CONNECT_FAILED) {
             return new RingException(sprintf('Connection timed out errCode=%s', $errCode));
         }
-        if ($statusCode === -2) {
+
+        if ($statusCode === SWOOLE_HTTP_CLIENT_ESTATUS_REQUEST_TIMEOUT) {
             return new RingException('Request timed out');
+        }
+
+        if ($statusCode === SWOOLE_HTTP_CLIENT_ESTATUS_SERVER_RESET) {
+            return new RingException('Server reset');
         }
 
         return true;