+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Fixed
 
+- [#534](https://github.com/hyperf-cloud/hyperf/pull/534) Fixed Guzzle HTTP Client does not handle the response status is equal to `-3`;
 - [#541](https://github.com/hyperf-cloud/hyperf/pull/541) Fixed bug grpc client cannot be set correctly.
 
 ## Deleted